change script folder and locations
